Windows port (#14)
* Windows port - work in progress

* Fix posix builds

* Another posix fix

* Move curl to a separate file, we will not use it on Windows

* Oops forgot the new file

* Oops we need PRIVATE

* Yes yes, std

* The callback

* CURL init

* Socket handle type

* Windows port is working

* Fix MacOS builds

* Run tests on MacOS

* Check http status code

* Enable Windows builds

* Try to fix windows builds

* DIR for debugging

* Try to find OpenSSL

* A better DIR

* Another attempt to make Windows CI work

* Use MSVC

* Select MSVC

* Do not build tests for Windows

* Upload artifacts

* Fetch pre-built openssl from our servers

* Try to run Windows tests

* Of course it didn't just work

* Check cross compiling

* Aaaah

* Remove debug logging

* Static linking seems to be working, but we can't build tests

* Built tests on Windows after all

* Trying again

* Actually build tests

* Hopefully final cleanup

* Handle redirects manually

+++ include/srtc/srtp_connection.h
@@ -36,11 +36,11 @@ public:
     bool unprotectIncomingControl(const ByteBuffer& packetData, ByteBuffer& output);
 
     // Implementation
-    SrtpConnection(const std::shared_ptr<SrtpCrypto>& crypto, bool isSetupActive, uint16_t profileId);
+    SrtpConnection(const std::shared_ptr<SrtpCrypto>& crypto, bool isSetupActive, unsigned long profileId);
 
 private:
     const std::shared_ptr<SrtpCrypto> mCrypto;
-    const uint16_t mProfileId;
+    const unsigned long mProfileId;
 
     struct ChannelKey {
         uint32_t ssrc;
